marketplace.app/requirements.txt
Flask
Flask-Assets
Flask-Mobility
Flask-Marshmallow
Flask-SQLalchemy
Flask-Script
Flask-Migrate
Flask-Session
Flask-Login
marshmallow
marshmallow-sqlalchemy
psycopg2-binary
Flask-RESTful
cssmin
jsmin
SQLAlchemy
SQLAlchemy-Searchable
flask_mail
python-dotenv
celery
redis
SQLAlchemy-Searchable
